Fixes certdir test.
Having modified the default for CRL checks in CANL to be less
stringent https://github.com/italiangrid/voms-api-java/blob/develop/src%2Fmain%2Fjava%2Forg%2Fitaliangrid%2Fvoms%2Futil%2FCertificateValidatorBuilder.java#L47
the test didn't fail as the CRL which revoked test0 was expired.

The test was changed to setup an empty certdir, so the failure
is now that no valid issuer credential is found for the test0
certificate.