Adapt the testsuite to work with the legacy clients
Add:
* a "legacy" tag to every test compatible with the legacy clients
* a `client_version` variable, which needs to be set to `2` to enable
  the legacy behavior

When the outputs of the two client flavors differ, an intermediate
variable called expected is initialized, depending on the value of
`client_version`.

To avoid a warning (and a non-zero return) from the legacy
voms-proxy-init when using -noregen, the proxies are created specifying
explicitly the duration.

Work in progress.